climb+push+XRToolKit
I added a script that allows the player to climb up the one net - can add the climbable script to other objects (possibly, needs to be play-tested). I added the model of the jungle gym. I added a script that allows the player to push the purple spiked objects (the script should be able to be attached to other objects as well). Added from Unity thru Project Manager the Installation of XR Interaction ToolKit that you can get from here: https://docs.unity3d.com/Packages/com.unity.xr.interaction.toolkit@3.0/manual/installation.html#:~:text=This%20package%20was%20not%20listed,unity.

Make sure to download the 2022.1 and later one of it. I am willing to help out with this! But it just made it ten times easier to do what we want it to do. Still needs to be play-tested but those are my changes!